Refactor GD32 timers and platform glue
Introduce a unified `gd32_timers` module for DWT delays, millis/uptime, SysTick, and timer startup, replacing the old split timer/delay/millis sources and wiring board init to `gd32::timers::Start()`. Migrate users (USB, hwclock, EMAC, timing wrappers) to the new `gd32::DelayUs()` / `gd32::Millis()` API, add shared GPIO macro/types in `gd32_gpio_macros.h`, and centralize optional subsystem debug macros in `gd32_debug.h`. The change also includes broad conditional/format cleanup, updates to logic-analyzer/status LED behavior, and soft UART/GPIO configuration refinements to align platform code paths.

+++ lib-gd32/device/usb/usb_vbus.cpp
@@ -30,13 +30,13 @@
 
 extern "C" {
 void usb_vbus_drive(uint8_t state) { // NOLINT
-    printf("usb_vbus_drive: %u\n", state);
+    printf("usb_vbus_drive: %u\n", static_cast<unsigned>(state));
 #if defined(USB_HOST_VBUS_GPIOx)
     if (0U == state) {
         gpio_bit_reset(USB_HOST_VBUS_GPIOx, USB_HOST_VBUS_GPIO_PINx);
     } else {
         gpio_bit_set(USB_HOST_VBUS_GPIOx, USB_HOST_VBUS_GPIO_PINx);
     }
-#endif
+#endif // USB_HOST_VBUS_GPIOx
 }
 }
